requirements.txt
annotated-types==0.7.0
gitdb==4.0.11
GitPython==3.1.41
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
setuptools==69.0.3
smmap==5.0.1
typing_extensions==4.12.2
wheel==0.43.0
